// src/types/form.d.ts
export interface FormDataInterface {
    id: string
    chalaniNumber: string
    date: string
    dartaNumber: string
    dartaMiti: string
    officeName: string
    fiscalYear: string
    asul: number
    aniyamit: number
    paperProof: number
    peski: number
    total: number
    chalaniNumber2: string
    chalaniDate: string
    ministry: string
    barsikPratibedan: string
    samparisayad_anurodh_rakam: number
    lagat_katta_ko_bibarad: string
    samparisad_huna_nasakeko: string
}